+++ apps/admin-x-activitypub/src/components/navigation/MainNavigation.tsx
